requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlmodel
sqlalchemy
python-dotenv
python-multipart
